Bounds: rename enrich to bounds; add integer range bounds of updates in postconditions; some small fixes in generated bounds
